requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
environ==1.0
flex==6.14.1
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpointer==2.2
jwcrypto==1.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==21.3
pycparser==2.20
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.16.0
rfc3987==1.3.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==4.1.1
urllib3==1.26.4
validate-email==1.3
wrapt==1.13.3
